Rename store/thunks/handlers to store/thunks/filetypes
